+++ libs/platform/src/lib/search-field/public_api.ts
@@ -1,6 +1,5 @@
 export * from './search-field.component';
 export * from './search-field.module';
 export * from './search-field-mobile/search-field/search-field-mobile.component';
-export * from './search-field-mobile/search-field-mobile.module';
 
 export * from './search-field-mobile/search-field-mobile.interface';