+++ libs/platform/src/lib/table/table.service.ts
@@ -2,6 +2,7 @@ import { EventEmitter, Injectable, Optional } from '@angular/core';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { DOWN_ARROW, LEFT_ARROW, RIGHT_ARROW, UP_ARROW } from '@angular/cdk/keycodes';
 import { skip } from 'rxjs/operators';
+import equal from 'fast-deep-equal';
 
 import { KeyUtil, RtlService } from '@fundamental-ngx/core/utils';
 
@@ -108,9 +109,11 @@ export class TableService {
         const newSortRules: CollectionSort[] = [...sortRules];
         const state: TableState = { ...prevState, sortBy: newSortRules };
 
-        this.setTableState(setCurrentPageToState(state, 1));
+        if (!equal(prevSortRules, state.sortBy)) {
+            this.setTableState(setCurrentPageToState(state, 1));
 
-        this.sortChange.emit({ current: state.sortBy, previous: prevSortRules });
+            this.sortChange.emit({ current: state.sortBy, previous: prevSortRules });
+        }
     }
 
     /** Add sort rules to the existing ones */
@@ -125,9 +128,11 @@ export class TableService {
 
         const state: TableState = { ...prevState, sortBy: newSortRules };
 
-        this.setTableState(setCurrentPageToState(state, 1));
+        if (!equal(prevSortRules, state.sortBy)) {
+            this.setTableState(setCurrentPageToState(state, 1));
 
-        this.sortChange.emit({ current: state.sortBy, previous: prevSortRules });
+            this.sortChange.emit({ current: state.sortBy, previous: prevSortRules });
+        }
     }
 
     /** Set new sort rules */
@@ -138,9 +143,11 @@ export class TableService {
         const newFilterRules: CollectionFilter[] = [...filterRules];
         const state: TableState = { ...prevState, filterBy: newFilterRules };
 
-        this.setTableState(setCurrentPageToState(state, 1));
+        if (!equal(prevFilterRules, state.filterBy)) {
+            this.setTableState(setCurrentPageToState(state, 1));
 
-        this.filterChange.emit({ current: state.filterBy, previous: prevFilterRules });
+            this.filterChange.emit({ current: state.filterBy, previous: prevFilterRules });
+        }
     }
 
     /** Add filter rules to the existing ones */
@@ -155,9 +162,11 @@ export class TableService {
 
         const state: TableState = { ...prevState, filterBy: newFilterRules };
 
-        this.setTableState(setCurrentPageToState(state, 1));
+        if (!equal(prevFilterRules, state.filterBy)) {
+            this.setTableState(setCurrentPageToState(state, 1));
 
-        this.filterChange.emit({ current: state.filterBy, previous: prevFilterRules });
+            this.filterChange.emit({ current: state.filterBy, previous: prevFilterRules });
+        }
     }
 
     /** Set group rules */
@@ -168,9 +177,11 @@ export class TableService {
         const newGroups: CollectionGroup[] = [...groups];
         const state: TableState = { ...prevState, groupBy: newGroups };
 
-        this.setTableState(state);
+        if (!equal(prevGroups, state.groupBy)) {
+            this.setTableState(state);
 
-        this.groupChange.emit({ current: state.groupBy, previous: prevGroups });
+            this.groupChange.emit({ current: state.groupBy, previous: prevGroups });
+        }
     }
 
     /** Add group rules to the existing ones */
@@ -185,9 +196,11 @@ export class TableService {
 
         const state: TableState = { ...prevState, groupBy: newGroups };
 
-        this.setTableState(state);
+        if (!equal(prevGroups, state.groupBy)) {
+            this.setTableState(state);
 
-        this.groupChange.emit({ current: state.groupBy, previous: prevGroups });
+            this.groupChange.emit({ current: state.groupBy, previous: prevGroups });
+        }
     }
 
     /** Freeze table to column */
@@ -207,9 +220,11 @@ export class TableService {
         const newColumns = [...columns];
         const state: TableState = { ...prevState, columns: newColumns };
 
-        this.setTableState(state);
+        if (!equal(prevColumns, state.columns)) {
+            this.setTableState(state);
 
-        this.columnsChange.emit({ current: state.columns, previous: prevColumns });
+            this.columnsChange.emit({ current: state.columns, previous: prevColumns });
+        }
     }
 
     /** Set page */